requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dash==3.0.2
dnspython==2.7.0
eventlet==0.39.1
Flask==3.0.3
Flask-SocketIO==5.5.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.34.1
nest-asyncio==1.6.0
packaging==24.2
plotly==6.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
retrying==1.3.4
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.21.0
